tests/e2e: separate coverage tests for exec commands
